move code to format.go
